+++ Task_1_1_1/app/src/main/java/sorter/Print.java
@@ -1,4 +1,4 @@
-package task_1_1_1;
+package sorter;
 
 /**
  *class for print somethings.
