+++ Task_1_1_1/app/src/main/java/sorter/Swap.java
@@ -1,4 +1,4 @@
-package task_1_1_1;
+package sorter;
 
 
 /**
