+++ Task_1_1_1/app/src/main/java/sorter/Timer.java
@@ -1,4 +1,4 @@
-package task_1_1_1;
+package sorter;
 
 import java.time.Duration;
 import java.time.Instant;
