+++ Task_1_1_1/app/src/test/java/sorter/HeapSortTest.java
@@ -1,15 +1,16 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package task_1_1_1;
-
-import task_1_1_1.HeapSort;
+package sorter;
 
 import java.util.Arrays;
 import java.util.Random;
 
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+
+import sorter.HeapSort;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class HeapSortTest {
